requirements.txt
discord.py>=2.0.0
wavelink>=2.4.0
python-dotenv